Pulled prompt runners & model code apart for readability
