+++ langdspy/__init__.py
@@ -1,7 +1,7 @@
 from .field_descriptors import InputField, OutputField, InputFieldList, HintField, OutputFieldEnum
 from .prompt_strategies import PromptSignature, PromptStrategy, DefaultPromptStrategy
-
-from .model import Model, PromptRunner, MultiPromptRunner
+from .prompt_runners import PromptRunner, RunnableConfig, Prediction
+from .model import Model, TrainedModelState
 
 from . import formatters
 from . import transformers
